:truck: Move CHART_ANNOTATIONS_CONFIG to own file
