+++ libs/designsystem/src/lib/components/chart/configs/type.config.ts
@@ -1,5 +1,4 @@
 import { ActiveElement, Chart, ChartEvent, ChartOptions } from 'chart.js';
-import { AnnotationTypeRegistry } from 'chartjs-plugin-annotation';
 
 import { ColorHelper, DesignTokenHelper } from '../../../helpers';
 import { ChartTypesConfig } from '../chart.types';
@@ -100,38 +99,6 @@ export const CHART_TYPES_CONFIG: ChartTypesConfig = {
         },
       },
     },
-  },
-};
-
-/* TODO: Find a proper place to store these two */
-/* The chart.js annotation does not allow for changing the 
-defaults it comes with. In order to have sensible defaults 
-this object is used instead and manually merged with the 
-rest of the annotations object */
-const borderColor = getThemeColorHexString('semi-dark');
-const borderDash: [number, number] = [6, 3];
-const borderWidth = 1;
-
-export const CHART_ANNOTATIONS_CONFIG: AnnotationTypeRegistry = {
-  line: {
-    borderDash,
-    borderWidth,
-    borderColor,
-  },
-  ellipse: {
-    borderDash,
-    borderWidth,
-    borderColor,
-    backgroundColor: 'transparent',
-  },
-  box: {
-    borderDash,
-    borderWidth,
-    borderColor,
-    backgroundColor: 'transparent',
-  },
-  point: {
-    backgroundColor: 'initial',
   },
 };
